Add wall interface, abstract class, and standard class. Add Level class.
Walls have been added but currently the do nothing other than visually creating a line.

The Level class is there to contain walls and other objects that will be added to the level. This will also allow for easy movement of all objects in the level at once.

A box has been put in the middle to help testing collisions in the future.